Refactor tokenization scripts: remove legacy files and consolidate functionality into a new model tokenizer with enhanced comparison features for different encodings.
