+++ src/a2/a1_encodings.py
